Added a components sub-folder to src
