--- src/Card/ICardProps.ts
@@ -1,6 +1,0 @@
-export interface ICardProps {
-  className?: string;
-  height?: string;
-  marginTop?: string;
-  width?: string;
-}
