+++ src/components/Card/ICardProps.ts
@@ -1,0 +1,30 @@
+/**
+ * Properties of the Card component
+ */
+export interface ICardProps {
+  /**
+   * ClassName of the Card
+   */
+  className?: string;
+
+  /**
+   * Height of the Card
+   *
+   * @default "auto"
+   */
+  height?: string;
+
+  /**
+   * Margin at the top of the Card
+   *
+   * @default "auto"
+   */
+  marginTop?: string;
+
+  /**
+   * Width of the Card
+   *
+   * @default "100%"
+   */
+  width?: string;
+}
