+++ tsconfig.json
@@ -1,20 +1,26 @@
 {
   "compilerOptions": {
-    "baseUrl": "src",
-    "skipLibCheck": true,
-    "target": "es5",
+    "sourceMap": true,
     "module": "commonjs",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "target": "es5",
     "jsx": "react",
-    "declaration": true,
-    "outDir": "build/dist",
-    "strict": true,
+    "allowJs": false,
     "moduleResolution": "node",
-    "typeRoots": ["node_modules/@types"],
+    "forceConsistentCasingInFileNames": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noImplicitAny": true,
+    "strictNullChecks": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "noUnusedLocals": true,
+    "declaration": true,
+    "declarationDir": "./build/dist/src",
+    "outDir": "./build/dist/src",
+    "rootDirs": ["src"],
     "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true
+    "skipLibCheck": true,
+    "downlevelIteration": true
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "build"]
 }
